pyproject.toml
[project]
name = "python-agentic-ai-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "httpx>=0.28.1",
    "ipython>=9.7.0",
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
